fix(native): move onboarding components out of app/ route tree
Expo Router treats every .tsx under app/ as a route. Keeping
shared components under app/(onboarding)/components/ polluted the
route registry (spurious /components/* entries) and broke typed
hrefs. Relocated to components/onboarding/ and updated imports to
the @/ alias.
