+++ apps/native/app/(onboarding)/difficulty-questions.tsx
@@ -7,9 +7,9 @@ import Svg, { Circle, ClipPath, Defs, G, Path, Rect } from "react-native-svg";
 import type { StudyDifficulty } from "@pruvi/shared";
 import { useOnboardingActions, useOnboardingStore } from "@/stores/onboardingStore";
 
-import { ExamCard } from "./components/exam-card";
-import { OnboardingHeader } from "./components/onboarding-header";
-import { PrimaryButton } from "./components/primary-button";
+import { ExamCard } from "@/components/onboarding/exam-card";
+import { OnboardingHeader } from "@/components/onboarding/onboarding-header";
+import { PrimaryButton } from "@/components/onboarding/primary-button";
 
 // --- Difficulty Icons ---
