+++ apps/native/components/onboarding/progress-bar.tsx
